+++ src/main/java/sopt/makers/authentication/adapter/in/web/controller/auth/AuthApiController.java
@@ -6,12 +6,12 @@ import sopt.makers.authentication.adapter.in.web.common.BaseResponse;
 import sopt.makers.authentication.adapter.in.web.common.code.AuthSuccess;
 import sopt.makers.authentication.adapter.in.web.dto.auth.request.AuthRequest;
 import sopt.makers.authentication.adapter.in.web.dto.auth.response.AuthResponse;
+import sopt.makers.authentication.adapter.in.web.util.CookieUtil;
+import sopt.makers.authentication.adapter.in.web.util.ResponseUtil;
 import sopt.makers.authentication.application.port.in.auth.AuthenticateSocialAccountUsecase;
 import sopt.makers.authentication.application.port.in.auth.CreatePhoneVerificationUsecase;
 import sopt.makers.authentication.application.port.in.auth.SignUpUsecase;
 import sopt.makers.authentication.application.port.in.auth.VerifyPhoneVerificationUsecase;
-import sopt.makers.authentication.support.util.CookieUtil;
-import sopt.makers.authentication.support.util.ResponseUtil;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
