+++ src/main/java/sopt/makers/authentication/adapter/in/web/controller/auth/SocialAccountApiController.java
@@ -4,9 +4,9 @@ import sopt.makers.authentication.adapter.in.web.common.BaseResponse;
 import sopt.makers.authentication.adapter.in.web.common.code.SocialAccountSuccess;
 import sopt.makers.authentication.adapter.in.web.dto.auth.request.SocialAccountRequest;
 import sopt.makers.authentication.adapter.in.web.dto.auth.response.AuthResponse;
+import sopt.makers.authentication.adapter.in.web.util.ResponseUtil;
 import sopt.makers.authentication.application.port.in.auth.GetSocialAccountUsecase;
 import sopt.makers.authentication.application.port.in.auth.UpdateSocialAccountUsecase;
-import sopt.makers.authentication.support.util.ResponseUtil;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
