+++ src/main/java/sopt/makers/authentication/adapter/in/web/controller/user/UserApiController.java
@@ -7,10 +7,10 @@ import sopt.makers.authentication.adapter.in.web.common.BaseResponse;
 import sopt.makers.authentication.adapter.in.web.common.code.UserSuccess;
 import sopt.makers.authentication.adapter.in.web.dto.user.request.UserRequest;
 import sopt.makers.authentication.adapter.in.web.dto.user.response.UserResponse;
+import sopt.makers.authentication.adapter.in.web.util.ResponseUtil;
 import sopt.makers.authentication.application.port.in.user.GetUserProfileUsecase;
 import sopt.makers.authentication.application.port.in.user.UpdateUserProfileUsecase;
 import sopt.makers.authentication.domain.user.Part;
-import sopt.makers.authentication.support.util.ResponseUtil;
 import sopt.makers.authentication.support.validator.UserIdValidator;
 import sopt.makers.authentication.support.validator.UserSearchConditionValidator;
