+++ src/main/java/sopt/makers/authentication/adapter/in/web/security/filter/JwtAuthenticationFilter.java
@@ -1,12 +1,12 @@
-package sopt.makers.authentication.support.security.filter;
+package sopt.makers.authentication.adapter.in.web.security.filter;
 
 import static sopt.makers.authentication.common.constant.SystemConstant.WHITELIST_WILDCARD;
 import static sopt.makers.authentication.domain.auth.exception.AuthFailure.MISSING_AUTHORIZATION_HEADER;
 
+import sopt.makers.authentication.adapter.in.web.security.authentication.ApiKeyAuthentication;
+import sopt.makers.authentication.adapter.in.web.security.authentication.CustomAuthentication;
 import sopt.makers.authentication.adapter.out.jwt.service.JwtAuthAccessTokenService;
 import sopt.makers.authentication.domain.auth.exception.AuthException;
-import sopt.makers.authentication.support.security.authentication.ApiKeyAuthentication;
-import sopt.makers.authentication.support.security.authentication.CustomAuthentication;
 
 import java.io.IOException;
